* Fixing bits and pieces

+++ SIN/SINRuntime/Src/SINDivisionOperator.cpp
@@ -27,8 +27,8 @@ namespace SIN {
 	
 	MemoryCell * DivisionOperator::ManageNumber(const MemoryCellNumber & mc1, const MemoryCell & mc2) {
 		if (mc2.Type() == MemoryCell::NUMBER_MCT) {
-			long long num1 = static_cast<long long>(mc1.GetValue());
-			long long num2 = static_cast<long long>(static_cast<const MemoryCellNumber &>(mc2).GetValue());
+			long num1 = static_cast<long>(mc1.GetValue());
+			long num2 = static_cast<long>(static_cast<const MemoryCellNumber &>(mc2).GetValue());
 			if(num2 != 0)
 				return SINEWCLASS(MemoryCellNumber, (num1/num2));
 			else
